backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
python-dotenv
langchain
langchain-experimental
langchain-groq
groq